+++ build.gradle
@@ -17,12 +17,13 @@ buildscript {
 
 repositories {
     mavenLocal()
+    mavenCentral()
     jcenter()
 }
 
 dependencies {
     compile group: 'com.google.guava', name: 'guava', version: '18.0'
-    
+
     compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.12'
     compile group: 'ch.qos.logback', name: 'logback-classic', version: '1.1.3'
 
@@ -36,8 +37,8 @@ jar {
     manifest {
         attributes(
             'Implementation-Version': "${version}",
-            'Implementation-Title': 'Bootenv - Java utility',
-            'Created-By': 'Bootenv',
+            'Implementation-Title': '>bootenv - Java',
+            'Created-By': '>bootenv',
             'Built-By': System.getProperty('user.name'),
             'Built-Date': new Date().format('yyyy-MM-dd HH:mm:ss')
         )
