+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'bootenv-java-utility'
+rootProject.name = 'bootenv'
