+++ src/main/resources/logback.xml
@@ -7,7 +7,7 @@
         </encoder>
     </appender>
 
-    <logger name="org.bootenv.utility" level="DEBUG"/>
+    <logger name="org.bootenv" level="DEBUG"/>
 
     <root level="WARN">
         <appender-ref ref="STDOUT"/>
